Create token interface, refactor stuff, and write regex's for each kind of token.
