+++ library/src/test/java/kwasm/format/text/token/util/NumTest.kt
@@ -12,16 +12,14 @@
  * limitations under the License.
  */
 
-package kwasm.format.text
+package kwasm.format.text.token.util
 
-import com.google.common.truth.Truth.assertAbout
 import com.google.common.truth.Truth.assertThat
 import kwasm.format.ParseException
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
-import java.util.concurrent.Executors.callable
 
 @RunWith(JUnit4::class)
 @UseExperimental(ExperimentalUnsignedTypes::class)
